requirements.txt
openai>=1.8.0
jsonformer>=0.12.0
langchain>=0.1.1
langchain-openai>=0.0.3
numpy==1.24.1
python-dotenv==1.0.0
tenacity==8.2.2
tiktoken>=0.5.2
transformers>=4.28.1
pandas>=1.5.3
scikit-learn>=1.4.0
wikipedia>=1.4.0
loguru>=0.7.2
datasets>=2.15.0
trl>=0.7.6
peft>=0.7.1
jsonlines>=4.0.0
torchmetrics>=1.2.1
sentencepiece==0.1.99
protobuf==4.25.1